Changing namespace Python module: `shelephant.cli` -> `shelephant`
