+++ shelephant/cli/shelephant_mv.py
@@ -26,8 +26,8 @@ import docopt
 import os
 
 from .. import __version__
-from . import ShelephantCopy
-from . import YamlGetItem
+from .. import ShelephantCopy
+from .. import YamlGetItem
 
 
 def main():
